[SERVER] remove listenserver, create server/main.cpp
This paths the way to remove the STANDALONE directive (on which the source is divided into client and server code)

the listenserver has been a feature to route calls from a local player to a local server without roundtripping through network sockets.

The benefit (**slightly** less delay) of that do not overweight the cons (complex code).